Time: 27 ms (82.47%), Space: 21.5 MB (99.07%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int minEatingSpeed(vector<int>& piles, int h) {
         int n = piles.size();
-        long long maxi = *max_element(piles.begin(),piles.end());
+        int maxi = *max_element(piles.begin(),piles.end());
         int low = 1;
-        long long high = maxi;
+        int high = maxi;
         while(low<=high){
             int mid = low+(high-low)/2;
             long long cnt = 0;
